requirements.txt
openai
langchain
pinecone-client
fastapi
PyPDF2
chardet
uvicorn
gunicorn